% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-load-cc-expr.r
\name{create_load_cc_expr}
\alias{create_load_cc_expr}
\title{Create an expression to load a Computational Component}
\usage{
create_load_cc_expr(load_cc_expr)
}
\arguments{
\item{load_cc_expr}{a string or expression that should be use to load 
the computational components.}
}
\description{
An expression to load a computational component can be either
a raw expression, a variable holding the expression, or a string. The
return is an unevaluated expression.
}
