% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-site.r
\name{ld_bundle_doc}
\alias{ld_bundle_doc}
\title{Create a `listdown` Document Bundle}
\usage{
ld_bundle_doc(cc, header, ld)
}
\arguments{
\item{cc}{the computational component list that will be presented.}

\item{header}{a `list` with the header information for the document.}

\item{ld}{a `listdown` object describing how to present the computational
components.}
}
\description{
A page bundle encapsulates the computational components, 
R Markdown header, and listdown object. Together, these three objects 
are sufficient to create a document, which can be written with the
`ld_create_document()` function.
}
\seealso{
ld_create_document
}
