% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk-options.r
\name{ld_chunk_opts}
\alias{ld_chunk_opts}
\title{Apply Chunk Options to a Presentation Object}
\usage{
ld_chunk_opts(pres_obj, chunk_name = NULL, ..., arg_list = NULL)
}
\arguments{
\item{pres_obj}{the presentation list element whose chunk options should
be modified.}

\item{chunk_name}{the name of the chunk. By default this is NULL, 
correpsonding to no chunk name.}

\item{...}{named chunk options and their values.}

\item{arg_list}{the list of chunk options can be specified. If ... 
parameters are specified, then this option is ignored.}
}
\description{
This function allows the user to set chunk options for 
individual elements of a presentation list.
}
