% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headers.r
\name{ld_ioslides_header}
\alias{ld_ioslides_header}
\title{Create an ioslides Presentation Header}
\usage{
ld_ioslides_header(title, author = NULL, date = NULL, theme = "flatly")
}
\arguments{
\item{title}{the title of the page.}

\item{author}{the author of the page. The default is NULL - no author.}

\item{date}{the date for the page. The default is NULL - no date.}

\item{theme}{the ioslides theme for the document. The default is "flatly".}
}
\description{
Output an ioslides R Markdown header with specified parameters.
}
