% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk-writer.r
\name{ld_make_chunks}
\alias{ld_make_chunks}
\title{Write a listdown Object to a String}
\usage{
ld_make_chunks(ld, rmd_dir)
}
\arguments{
\item{ld}{the listdown object that provides
information on how a presentation object should be displayed in the
output.}

\item{rmd_dir}{the R Markdown directory.}
}
\description{
After a presentation list and listdown object have been
constructed the chunks can be rendered to a string, which can be appended
to a file, with appropriate headers, resulting in a compilable R Markdown
document.
}
\seealso{
\code{\link{listdown}}
}
