% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headers.r
\name{ld_rmarkdown_header}
\alias{ld_rmarkdown_header}
\title{Create an R Markdown Header}
\usage{
ld_rmarkdown_header(
  title,
  author = NULL,
  date = NULL,
  output = c("html_document", "pdf_document", "word_document")
)
}
\arguments{
\item{title}{the title of the page.}

\item{author}{the author of the page. The default is NULL - no author.}

\item{date}{the date for the page. The default is NULL - no date.}

\item{output}{the output format of the page. If NULL then no output format.
The default is an html document.}
}
\description{
Output an R Markdown header with specified parameters.
}
