% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-site.r
\name{ld_site_yaml}
\alias{ld_site_yaml}
\title{Create a Minimalist Site YAML List}
\usage{
ld_site_yaml(site_name, tab_name, rmd_name, navbar_title = site_name)
}
\arguments{
\item{site_name}{the name of the site.}

\item{tab_name}{the name of the tabs on the site.}

\item{rmd_name}{the name of the Rmarkdown files that will generate the
respective tabs.}

\item{navbar_title}{the title of the navigation bar (Default is the
`site_name` argument.}
}
\description{
Create a Minimalist Site YAML List
}
