% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headers.r
\name{ld_workflowr_header}
\alias{ld_workflowr_header}
\title{Create a workflowr Header}
\usage{
ld_workflowr_header(title, toc = FALSE)
}
\arguments{
\item{title}{the title of the page.}

\item{toc}{should the table of contents be generated? Default FALSE.}
}
\value{
An S3 instance of class `listdown_header` containing the 
R Markdown `workflowr` header data.
}
\description{
Output a workflowr R Markdown header with specified title.
}
