% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld-write-file.r
\name{ld_write_file}
\alias{ld_write_file}
\title{Write to an R Markdown File}
\usage{
ld_write_file(rmd_header, ld, file_name)
}
\arguments{
\item{rmd_header}{either a character or listdown_header with R Markdown
header information.}

\item{ld}{the listdown object that provides
information on how a presentation object should be displayed in the
output.}

\item{file_name}{the output file to write to.}
}
\description{
This function takes header information and a listdown 
object and writes to a specified file.
}
\examples{
\donttest{
library(ggplot2)

cc_list <- list(
  Linear = ggplot(anscombe, aes(x = x1, y = y1)) + geom_point(),
 `Non Linear` = ggplot(anscombe, aes(x = x2, y = y2)) + geom_point(),
 `Outlier Vertical`= ggplot(anscombe, aes(x = x3, y = y3)) + geom_point(),
 `Outlier Horizontal` =  ggplot(anscombe, aes(x = x4, y = y4)) +
   geom_point())

saveRDS(cc_list, "cc-list.rds")

ld <- listdown(readRDS("cc-list.rds"),
               package = "ggplot2")

ld_write_file(ld_rmarkdown_header(title = "The Anscombe Quartet",
                                  author = "Francis Anscombe",
                                  date = "1973"),
              ld,
              "anscombe-quartet.rmd")
}
}
