% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld-write-file.r
\name{ld_write_file}
\alias{ld_write_file}
\title{Write to an R Markdown File}
\usage{
ld_write_file(rmd_header, ld, file_name)
}
\arguments{
\item{rmd_header}{either a character or listdown_header with R Markdown
header information.}

\item{ld}{the listdown object that provides
information on how a presentation object should be displayed in the
output.}

\item{file_name}{the output file to write to.}
}
\description{
This function takes header information and a listdown 
object and writes to a specified file.
}
