% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk-writer.r
\name{listdown}
\alias{listdown}
\title{Create a listdown Object}
\usage{
listdown(
  load_cc_expr,
  package = NULL,
  decorator = list(),
  init_expr = NULL,
  default_decorator = identity,
  ...
)
}
\arguments{
\item{load_cc_expr}{an unquoted expression to load the presentation list.}

\item{package}{a quoted list of package required by the outputted document.}

\item{decorator}{a named list mapping the potential types of list elements
to a decorator function.}

\item{init_expr}{an initial expression that will be added to the outputted
document after the libraries have been called.}

\item{default_decorator}{the decorator to use for list elements whos type
is not inherited from the decorator list. If NULL then the those
elements will not be included when the chunks are written. By default
this is identity, meaning that the elements will be passed directly 
(through the identity() function).}

\item{...}{default options sent to the chunks of the outputted document.}
}
\description{
A listdown object provides information for how a presentation
list should be used to create an R Markdown document. It requires an
unquoted expression indicating how the presentation list will be loaded.
In addition, libraries required by the outputted document and other
paraeters can be specified.
}
